+++ CHANGELOG.md
@@ -38,6 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Several dozen internal code improvements
 
 ### Fixed
+- Fixed IMDb ratings using a comma in localized number formats
 - Fixed toasts not displaying above some sheets
 - Fixed calendar/queue not updating instances when changed
 - Fixed sidebar colors on macOS 27